fix: Fix problemas in the controller and service Login. Change in the migratio and model Produt the price DOBLE to DECIMAL

+++ back-end/src/database/models/Products.model.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: DataTypes.STRING,
-    price: DataTypes.DOUBLE,
+    price: DataTypes.DECIMAL,
     urlImage: DataTypes.STRING,
   },
   {
